refactor: extract ABC parser into js/activity/abc-parser.js
Move _adjustPitch, _abcToStandardValue, _createPitchBlocks,
_searchIndexForMusicBlock, and parseABC out of activity.js and into
a dedicated js/activity/abc-parser.js module.

The new module follows the same AMD + guarded module.exports pattern
as js/activity/recorder.js. setupActivityAbcParser(activityInstance)
attaches this.parseABC to the activity instance, preserving the
existing import workflow and public API without any UX changes.

loader.js is updated with a path alias and a dep entry for
activity/abc-parser so it is loaded before activity/activity.

The activity_startup_recovery test sandbox is updated with a no-op
setupActivityAbcParser so the vm slice of activity.js continues to
evaluate without a ReferenceError.

Additionally, add a unit test suite under js/activity/__tests__/abc-parser.test.js
covering key scenarios to guarantee exact behavioral preservation.

Verified: npm run lint (0 new errors), prettier --check (pass),
npm test (5449/5449 pass).
